Refactor rename `cli` -> `cli_github_actions`
