+++ tools/cli_github_actions/handlers/github_actions.py
@@ -3,7 +3,7 @@ import re
 import uuid
 from typing import Any, Literal
 
-from tools.cli.globals import VERSION_PATTERN
+from tools.cli_github_actions.globals import VERSION_PATTERN
 
 def set_step_output(name: str, value: str) -> None:
     with open(os.environ["GITHUB_OUTPUT"], "a") as env:
